+++ packages/base-controller/src/BaseControllerV2.test.ts
@@ -2,7 +2,6 @@
 import type { Draft, Patch } from 'immer';
 import * as sinon from 'sinon';
 
-import { TestController } from './BaseControllerV1.test';
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
@@ -187,14 +186,8 @@ describe('isBaseController', () => {
     expect(isBaseController(controller)).toBe(true);
   });
 
-  it('should return false if passed a V1 controller', () => {
-    const controller = new TestController();
-    expect(isBaseController(controller)).toBe(false);
-  });
-
   it('should return false if passed a non-controller', () => {
     const notController = new JsonRpcEngine();
-    // @ts-expect-error Intentionally passing invalid input to test runtime behavior
     expect(isBaseController(notController)).toBe(false);
   });
 });